% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_master_to_wd.R
\name{copy_master_to_wd}
\alias{copy_master_to_wd}
\title{Copy Master_Template.xlsx to computer}
\usage{
copy_master_to_wd()
}
\description{
The Master_Template.xlsx file that is bundled with the \code{inportXML} package gets installed (by default) with your package.
This is often in a location the typical user never looks. \code{copy_master_to_wd} allows the user to copy this file automatically
without having to know where it was installed.
}
\examples{
#'simply run the function to copy the file to your working directory

}
