% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEntityXML.R
\name{createEntityXML}
\alias{createEntityXML}
\title{Create XMl for an InPort Entity from the InportXML metadata template spreadsheet.}
\usage{
createEntityXML(inFile, outFile)
}
\arguments{
\item{inFile}{The full path to the metadata template (e.g. "~/Master_Template.xlsx").}

\item{outFile}{The name of the resulting XML file (e.g. "Entity1.xml").}
}
\value{
An XML file for an InPort Entity.
}
\description{
For an InPort metadata rubric score of 100 \%, the submission of both an Item and at
least one Entity are required. This function will create the Entity XML. The term "Entity"
refers to specific metadata related to the structure of the Data Set, and must include
at least one "attribute" for successful XML upload. These instructions are further
explained in the "Master_Template.xlsx" file.
}
\examples{
#Be sure that inFile refers to the full path of the metadata template
createEntityXML(inFile = "~/Master_Template.xlsx", outFile = "Data_Entity1.xml")
}
