% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInPortXML.R
\name{createInPortXML}
\alias{createInPortXML}
\title{Create all XML files for direct upload of metadata to InPort.}
\usage{
createInPortXML(inFile, outFile)
}
\arguments{
\item{inFile}{The full path to the metadata template (e.g. "~/Master_Template.xlsx").}

\item{outFile}{An identifying name for output files (e.g. "CHL_metadata_V1.xml".)}
}
\value{
Two XML files; one for the Item XML and one for Entity XML. The filenames
will be concatenated with "master_" and "entity_" in a given directory.
}
\description{
This function calls both \code{\link{createItemXML}} and \code{\link{createEntityXML}} functions to
generate two separate XML files. As InPort is currently configured, the two XML files
must be uploaded in separate steps, and both are necessary for successful upload.
}
\examples{
#Be sure that inFile refers to the full path of the metadata template
createInPortXML(inFile = "~/Master_Template.xlsx", outFile = "X_metadata_V1.xml")

}
