% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createItemXML.R
\name{createItemXML}
\alias{createItemXML}
\title{Create XMl for an InPort data item from the InportXML metadata template spreadsheet.}
\usage{
createItemXML(inFile, outFile)
}
\arguments{
\item{inFile}{The full path to the metadata template (e.g. "~/Master_Template.xlsx").}

\item{outFile}{The name of the resulting XML file (e.g. "DataSet1.xml").}
}
\value{
An XML file for an InPort Item.
}
\description{
For an InPort metadata rubric score of 100%, the submission of both an Item and its
respective Entity are required. This function will create Item XML. Currently, the
term "Item" refers solely to a Data Set.
}
\examples{
#Be sure that inFile refers to the full path of the metadata template
createItemXML(inFile = "~/Master_Template.xlsx", outFile = "DataSet1.xml")

}
