% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inportxml.R
\docType{package}
\name{inportxml}
\alias{inportxml}
\title{inportxml: xml creation for Inport}
\description{
Tools to aid in creation of xml files for direct upload into Inport (\url{https://inport.nmfs.noaa.gov/inport/})
}
\details{
\itemize{
\item User must complete the Master xml template
\item This is then converted into an xml file(s) for user to upload into Inport
}

To learn more about using \code{inportxml}, start with the vignette: \code{browseVignettes(package="inportxml")} or click the index below
}
