% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openMasterTemplate.R
\name{openMasterTemplate}
\alias{openMasterTemplate}
\title{Open the master InPort template from session}
\usage{
openMasterTemplate(path_to_excel)
}
\arguments{
\item{path_to_excel}{Enter the full file path to Microsoft Excel on your Windows machine}
}
\description{
Calling this function will open the master InPort template that is packaged with InPortXML.
Only Windows systems are currently supported.
}
\examples{
openMasterTemplate(path_to_excel = "C:/Program Files (x86)/Microsoft Office/Office15/Excel.exe")

}
